@@ -7,7 +7,10 @@
     "start": "tsc",
     "tsc": "tsc",
     "tsc:w": "tsc -w",
-    "postinstall": "typings install"
+    "postinstall": "typings install",
+    "test": "karma start karma.conf.js",
+    "posttest": "node_modules/.bin/remap-istanbul -i coverage/coverage-final.json -o coverage -t html",
+    "coverage": "http-server -c-1 -o -p 9875 ./coverage"
   },
   "repository": {
     "type": "git",
@@ -19,13 +22,35 @@
     "url": "https://github.com/elasticrash/Angular2.leaflet.component/issues"
   },
   "dependencies": {
-    "@angular/common": "^2.3.0",
-    "@angular/core": "^2.3.0",
-    "@angular/http": "^2.3.0",
-    "@angular/platform-browser": "^2.3.0",
+    "@angular/common": "^2.4.1",
+    "@angular/core": "^2.4.1",
+    "@angular/http": "^2.4.1",
+    "@angular/platform-browser": "^2.4.1",
+    "@angular/platform-browser-dynamic": "^2.4.1",
     "leaflet": "^1.0.1",
-    "rxjs": "5.0.0-rc.4",
-    "zone.js": "^0.7.2"
+    "reflect-metadata": "^0.1.9",
+    "ts-helpers": "^1.1.1",
+    "rxjs": "5.0.1",
+    "zone.js": "^0.7.4"
   },
-  "homepage": "https://github.com/elasticrash/Angular2.leaflet.component#readme"
+  "homepage": "https://github.com/elasticrash/Angular2.leaflet.component#readme",
+  "devDependencies": {
+    "@angular/compiler-cli": "^2.3.1",
+    "@types/jasmine": "2.5.38",
+    "@types/node": "^6.0.42",
+    "angular-cli": "1.0.0-beta.24",
+    "codelyzer": "~2.0.0-beta.1",
+    "jasmine-core": "2.5.2",
+    "jasmine-spec-reporter": "2.5.0",
+    "karma": "1.2.0",
+    "karma-chrome-launcher": "^2.0.0",
+    "karma-cli": "^1.0.1",
+    "karma-jasmine": "^1.0.2",
+    "karma-remap-istanbul": "^0.2.1",
+    "karma-webpack": "^1.8.1",
+    "protractor": "~4.0.13",
+    "ts-node": "1.2.1",
+    "tslint": "^4.0.2",
+    "typescript": "~2.0.3"
+  }
 }
